Support multiple spatial fields with different names.
Previously, the list of spatially indexed fields was stored on the
Document module, rather than on the class that included the module. It
consequently got clobbered every time Mongoid::Spacial::Document was
included in a class.

Only the last model to be loaded would therefore be able to access it's
spatial fields properly.